requirements.txt
python-dotenv==1.0.1
pandas==2.2.2
requests==2.32.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
tenacity==9.0.0
plotly==5.18.0
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
numpy>=1.26.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
xlsxwriter>=3.1.0
us>=2.0.0
pyarrow>=14.0.0
tqdm>=4.66.0
pytest>=7.4.0
ruff>=0.1.0